+++ app/src/main/java/scot/raven/titanpad/settings/repository/SettingsRepository.kt
@@ -11,9 +11,9 @@ interface SettingsRepository {
     suspend fun updateSettings(settings: ApplicationSettings)
     suspend fun validateAndUpdateSettings(settings: ApplicationSettings): ApplicationSettings.ValidationResult
 
-    suspend fun exportSettings(settings: ApplicationSettings) : String
-    suspend fun exportSettings(configId: String, usageConfig: UsageConfig) : String
-    suspend fun exportSettingsWithoutAppData(configId: String, usageConfig: UsageConfig): String
+    fun exportSettings(settings: ApplicationSettings) : String
+    fun exportSettings(configId: String, usageConfig: UsageConfig) : String
+    fun exportSettingsWithoutAppData(configId: String, usageConfig: UsageConfig): String
     suspend fun importSettings(jsonData: String) : Boolean
     suspend fun importSettings(configId: String, jsonData: String) : Boolean
     suspend fun importSettingsWithoutAppData(configId: String, jsonData: String): Boolean